Enhance the project to add comments

+++ Socialcademy/Socialcademy/View Models/PostsViewModel.swift
@@ -46,7 +46,7 @@ class PostsViewModel: ObservableObject {
     }
     
     func makeNewPostViewModel() -> FormViewModel<Post> {
-        return FormViewModel(
+        return FormViewModel<Post>(
             initialValue: Post(title: "", content: "", author: postsRepository.user),
             action: { [weak self] post in
                 try await self?.postsRepository.create(post)
@@ -56,20 +56,21 @@ class PostsViewModel: ObservableObject {
     }
     
     func makePostRowViewModel(for post: Post) -> PostRowViewModel {
-        let deleteAction = { [weak self] in
-            try await self?.postsRepository.delete(post)
-            self?.posts.value?.removeAll { $0 == post }
-        }
-        let favoriteAction = { [weak self] in
-            let newValue = !post.isFavorite
-            try await newValue ? self?.postsRepository.favorite(post) : self?.postsRepository.unfavorite(post)
-            guard let i = self?.posts.value?.firstIndex(of: post) else { return }
-            self?.posts.value?[i].isFavorite = newValue
-        }
         return PostRowViewModel(
             post: post,
-            deleteAction: postsRepository.canDelete(post) ? deleteAction : nil,
-            favoriteAction: favoriteAction
+            deleteAction: {
+                guard postsRepository.canDelete(post) else { return nil }
+                return { [weak self] in
+                    try await self?.postsRepository.delete(post)
+                    self?.posts.value?.removeAll { $0 == post }
+                }
+            }(),
+            favoriteAction: { [weak self] in
+                let newValue = !post.isFavorite
+                try await newValue ? self?.postsRepository.favorite(post) : self?.postsRepository.unfavorite(post)
+                guard let i = self?.posts.value?.firstIndex(of: post) else { return }
+                self?.posts.value?[i].isFavorite = newValue
+            }
         )
     }
 }
